Merge curatedMetagenomicData List

To merge the list elements returned from curatedMetagenomicData into a single SummarizedExperiment or TreeSummarizedExperiment object, users will use mergeData() provided elements are the same dataType (see returnSamples). This is useful for analysis across entire studies (e.g. meta-analysis); however, when doing analysis across individual samples (e.g. mega-analysis) returnSamples is preferable.

mergeData(mergeList)

Arguments

mergeList
a list returned from curatedMetagenomicData where all of the elements are of the same dataType (see returnSamples)

Value

when mergeList elements are of dataType (see returnSamples) relative_abundance, a TreeSummarizedExperiment object is returned; otherwise, a SummarizedExperiment object is returned

Details

Internally, mergeData() must full join assays and rowData slots of each SummarizedExperiment or TreeSummarizedExperiment object (colData is merged slightly more efficiently by row binding). While dplyr methods are used for maximum efficiency, users should be aware that memory requirements can be large when merging many list elements.

See also

Examples

curatedMetagenomicData("LiJ_20.+.marker_abundance", dryrun = FALSE) %>%
    mergeData()
#> snapshotDate(): 2021-05-05
#> class: SummarizedExperiment 
#> dim: 77712 456 
#> metadata(0):
#> assays(1): marker_abundance
#> rownames(77712): 39491__A0A395UVM9__DXB76_04540
#>   39491__A0A395UZL1__DXB76_05950 ... 1262937__R7LHU6__BN805_01557
#>   712117__F3PBR4__HMPREF9056_02502
#> rowData names(0):
#> colnames(456): DLF005-IE DLM006-IE ... nHM612836 nHMX11726
#> colData names(28): study_name subject_id ... location smoker

curatedMetagenomicData("LiJ_20.+.pathway_abundance", dryrun = FALSE) %>%
    mergeData()
#> snapshotDate(): 2021-05-05
#> class: SummarizedExperiment 
#> dim: 27110 456 
#> metadata(0):
#> assays(1): pathway_abundance
#> rownames(27110): UNMAPPED UNINTEGRATED ... PWY-7539:
#>   6-hydroxymethyl-dihydropterin diphosphate biosynthesis III
#>   (Chlamydia)|g__Prevotella.s__Prevotella_buccae LACTOSECAT-PWY:
#>   lactose and galactose degradation
#>   I|g__Escherichia.s__Escherichia_coli
#> rowData names(0):
#> colnames(456): DLF005-IE DLM006-IE ... nHM612836 nHMX11726
#> colData names(28): study_name subject_id ... location smoker

curatedMetagenomicData("LiJ_20.+.relative_abundance", dryrun = FALSE) %>%
    mergeData()
#> snapshotDate(): 2021-05-05
#> class: TreeSummarizedExperiment 
#> dim: 688 456 
#> metadata(0):
#> assays(1): relative_abundance
#> rownames(688):
#>   k__Bacteria|p__Bacteroidetes|c__Bacteroidia|o__Bacteroidales|f__Bacteroidaceae|g__Bacteroides|s__Bacteroides_plebeius
#>   k__Bacteria|p__Bacteroidetes|c__Bacteroidia|o__Bacteroidales|f__Bacteroidaceae|g__Bacteroides|s__Bacteroides_caccae
#>   ...
#>   k__Bacteria|p__Proteobacteria|c__Betaproteobacteria|o__Neisseriales|f__Neisseriaceae|g__Neisseria|s__Neisseria_elongata
#>   k__Bacteria|p__Proteobacteria|c__Gammaproteobacteria|o__Enterobacterales|f__Morganellaceae|g__Providencia|s__Providencia_alcalifaciens
#> rowData names(7): kingdom phylum ... genus species
#> colnames(456): DLF005-IE DLM006-IE ... nHM612836 nHMX11726
#> colData names(28): study_name subject_id ... location smoker
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):
#> rowLinks: a LinkDataFrame (688 rows)
#> rowTree: 1 phylo tree(s) (10430 leaves)
#> colLinks: NULL
#> colTree: NULL