Return Samples Across Studies

To return samples across studies, users will use returnSamples() along with the sampleMetadata data.frame subset to include only desired samples and metadata. The subset sampleMetadata data.frame will be used to get the desired resources, mergeData will be used to merge them, and the subset sampleMetadata data.frame will be used again to subset the SummarizedExperiment or TreeSummarizedExperiment object to include only desired samples and metadata.

returnSamples(sampleMetadata, dataType, counts = FALSE)

Arguments

sampleMetadata
the sampleMetadata data.frame subset to include only desired samples and metadata

dataType
the data type to be returned; one of the following:

"gene_families"

"marker_abundance"

"marker_presence"

"pathway_abundance"

"pathway_coverage"

"relative_abundance"

counts
if FALSE (the default), relative abundance proportions are returned; if TRUE, relative abundance proportions are multiplied by read depth and rounded to the nearest integer prior to being returned

Value

when dataType = "relative_abundance", a TreeSummarizedExperiment object is returned; otherwise, a SummarizedExperiment object is returned

Details

At present, curatedMetagenomicData resources exists only as entire studies which requires potentially getting many resources for a limited number of samples. Furthermore, because it is necessary to use mergeData internally, the same caveats detailed under Details in mergeData apply here.

Examples

sampleMetadata %>%
    dplyr::filter(age >= 18) %>%
    dplyr::filter(!base::is.na(alcohol)) %>%
    dplyr::filter(body_site == "stool") %>%
    dplyr::select(where(~ !base::all(base::is.na(.x)))) %>%
    returnSamples("relative_abundance")

#> snapshotDate(): 2021-05-05

#> class: TreeSummarizedExperiment 
#> dim: 833 702 
#> metadata(0):
#> assays(1): relative_abundance
#> rownames(833):
#>   k__Bacteria|p__Bacteroidetes|c__Bacteroidia|o__Bacteroidales|f__Prevotellaceae|g__Prevotella|s__Prevotella_copri
#>   k__Bacteria|p__Bacteroidetes|c__Bacteroidia|o__Bacteroidales|f__Prevotellaceae|g__Prevotella|s__Prevotella_sp_CAG_520
#>   ...
#>   k__Bacteria|p__Actinobacteria|c__Actinobacteria|o__Corynebacteriales|f__Corynebacteriaceae|g__Corynebacterium|s__Corynebacterium_aurimucosum
#>   k__Bacteria|p__Actinobacteria|c__Actinobacteria|o__Corynebacteriales|f__Corynebacteriaceae|g__Corynebacterium|s__Corynebacterium_coyleae
#> rowData names(7): kingdom phylum ... genus species
#> colnames(702): JAS_1 JAS_10 ... YSZC12003_37879 YSZC12003_37880
#> colData names(36): study_name subject_id ... systolic_p zigosity
#> reducedDimNames(0):
#> mainExpName: NULL
#> altExpNames(0):
#> rowLinks: a LinkDataFrame (833 rows)
#> rowTree: 1 phylo tree(s) (10430 leaves)
#> colLinks: NULL
#> colTree: NULL